mostly completed unit tests
